backend/requirements.txt
geoip2==2.5.0
PyJWT==1.5.3
python-jose==1.4.0
